+++ app/OpenChat.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App;
+
+use Movim\Model;
+
+class OpenChat extends Model
+{
+    public $incrementing = false;
+    protected $primaryKey = ['user_id', 'jid'];
+    protected $fillable = ['jid'];
+
+    public function user()
+    {
+        return $this->belongsTo('App\User');
+    }
+}