+++ CHANGELOG.md
@@ -24,6 +24,7 @@ v0.26.1 (master)
 * Move the Movim configuration to a new movim:configuration node, use XEP-0004: Data Form to handle the data
 * Add a missing cascade foreign key between the sessions and users table
 * Remove the VisioLink and Onboarding popup and related code
+* Remove the Cache model and related table, move the remaining data in the users and open_chats table
 
 v0.26
 ---------------------------